requirements.txt
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.2
fastapi==0.116.1
filelock==3.18.0
frozenlist==1.7.0
greenlet==3.2.3
h11==0.16.0
hf-xet==1.1.7
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.3
idna==3.10
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain
langchain-community
langchain-core
langchain-huggingface
langchain-text-splitters
langsmith
MarkupSafe==3.0.2
multidict==6.6.3
numpy==1.26.4
orjson==3.11.1
packaging==25.0
pydantic==2.11.7
pydantic-settings==2.10.1
pypdf==5.9.0
python-dotenv==1.1.1
python-multipart==0.0.20
requests==2.32.4
requests-toolbelt==1.0.0
safetensors==0.6.1
sentence-transformers==5.0.0
setuptools==80.9.0
sniffio==1.3.1
starlette==0.47.2
tenacity==8.2.3
threadpoolctl==3.6.0
tokenizers==0.21.4
torch==2.7.1
tqdm==4.67.1
transformers==4.55.0
uvicorn==0.35.0
yarl==1.20.1
zstandard==0.23.0
